LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-transform-dynamic-expected.txt


FAIL Initially, the sibling-index() is 3 for #target assert_equals: expected "matrix(1, 0, 0, 1, 30, 0)" but got "none"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "matrix(1, 0, 0, 1, 20, 0)" but got "none"
